Refactor sorting for students
