+++ src/main/java/tfifteenfour/clipboard/logic/parser/SortCommandParser.java
@@ -1,15 +1,23 @@
 package tfifteenfour.clipboard.logic.parser;
 
-import java.util.Comparator;
-
-import tfifteenfour.clipboard.logic.commands.studentcommands.SortCommand;
+import tfifteenfour.clipboard.logic.CurrentSelection;
+import tfifteenfour.clipboard.logic.PageType;
+import tfifteenfour.clipboard.logic.commands.exceptions.CommandException;
+import tfifteenfour.clipboard.logic.commands.sortcommand.SortCommand;
+import tfifteenfour.clipboard.logic.commands.sortcommand.studentcomparators.AlphaNumericSidComparator;
+import tfifteenfour.clipboard.logic.commands.sortcommand.studentcomparators.AlphabeticalNameComparator;
 import tfifteenfour.clipboard.logic.parser.exceptions.ParseException;
-import tfifteenfour.clipboard.model.student.Student;
 
 /**
  * Parses input arguments and creates a new SortCommand object.
  */
 public class SortCommandParser implements Parser<SortCommand> {
+
+    private final CurrentSelection currentSelection;
+
+    public SortCommandParser(CurrentSelection currentSelection) {
+        this.currentSelection = currentSelection;
+    }
 
     /**
      * Parses the given {@code String} of arguments in the context of the SortCommand
@@ -19,38 +27,47 @@ public class SortCommandParser implements Parser<SortCommand> {
      * @return The new SortCommand object.
      * @throws ParseException If the user input does not conform to the expected format.
      */
-    public SortCommand parse(String args) throws ParseException {
+    public SortCommand parse(String args) throws ParseException, CommandException {
         String trimmedArgs = args.trim();
 
         SortCategory category = SortCategory.fromString(trimmedArgs);
 
+        if (currentSelection.getCurrentPage() != PageType.STUDENT_PAGE) {
+            throw new CommandException("Wrong page. Navigate to student page to sort students");
+        }
+
         switch (category) {
         case SORT_BY_NAME:
-            return new SortCommand(new AlphaNameComparator(), category.getCategory());
+            return new SortCommand(new AlphabeticalNameComparator(), category.getCategory());
         case SORT_BY_STUDENT_ID:
-            return new SortCommand(new NumericAlphaStudentIdComparator(), category.getCategory());
+            return new SortCommand(new AlphaNumericSidComparator(), category.getCategory());
         default:
             throw new ParseException("Unable to parse category for sorting");
         }
     }
 }
 
-/**
- * Comparator for sorting students by name.
- */
-class AlphaNameComparator implements Comparator<Student> {
-    @Override
-    public int compare(Student s1, Student s2) {
-        return s1.getName().toString().compareTo(s2.getName().toString());
-    }
-}
+enum SortCategory {
+    SORT_BY_NAME("name"),
+    SORT_BY_STUDENT_ID("id");
 
-/**
- * Comparator for sorting students by student ID.
- */
-class NumericAlphaStudentIdComparator implements Comparator<Student> {
-    @Override
-    public int compare(Student s1, Student s2) {
-        return s1.getStudentId().toString().compareTo(s2.getStudentId().toString());
+    private String category;
+
+    private SortCategory(String category) {
+        this.category = category;
+    }
+
+    public String getCategory() {
+        return this.category;
+    }
+
+    public static SortCategory fromString(String category) throws ParseException {
+        for (SortCategory sc : SortCategory.values()) {
+            if (sc.getCategory().equalsIgnoreCase(category)) {
+                return sc;
+            }
+        }
+
+        throw new ParseException("Unrecognised category for sort: " + category);
     }
 }
